Implement a more clear distinction between warnings and errors
Errors include faults that will cause a runtime error of some kind when
the Python script is executed. Warnings cover simpler mistakes such as
unused variables and unused imports.

This change also has the positive side effect that it makes the plugin
work better with most (if not all) PyFlakes versions, as it no longer
directly references attributes inside the messages module.